refactor: update public facades and model imports across services
