+++ api/specs/web-server/_announcements.py
@@ -7,7 +7,7 @@
 from fastapi import APIRouter
 from models_library.generics import Envelope
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.announcements._handlers import Announcement
+from simcore_service_webserver.announcements.models import Announcement
 
 router = APIRouter(
     prefix=f"/{API_VTAG}",
