+++ services/web/server/src/simcore_service_webserver/announcements/announcements_service.py
@@ -1,13 +1,9 @@
 """Announcements public facade per DESIGN.md §133-152."""
 
-# Models
 # Functions
 from ._api import list_announcements
-from ._models import Announcement
 
 __all__: tuple[str, ...] = (
-    # Models
-    "Announcement",
-    # Functions
+    # functions
     "list_announcements",
 )  # nopycln: file
