+++ services/web/server/src/simcore_service_webserver/announcements/models.py
@@ -1,0 +1,8 @@
+"""Announcements public facade for models per DESIGN.md §133-152."""
+
+from ._models import Announcement
+
+__all__: tuple[str, ...] = (
+    # models
+    "Announcement",
+)
