+++ services/web/server/src/simcore_service_webserver/catalog/catalog_service.py
@@ -6,11 +6,10 @@ from ._catalog_rest_client_service import (
     is_catalog_service_responsive,
     to_backend_service,
 )
-from ._models import ServiceKeyVersionDict
 from ._service import batch_get_my_services
 
 __all__: tuple[str, ...] = (
-    "ServiceKeyVersionDict",
+    # functions
     "batch_get_my_services",
     "get_service",
     "get_service_access_rights",
@@ -18,5 +17,4 @@ __all__: tuple[str, ...] = (
     "get_services_for_user_in_product",
     "is_catalog_service_responsive",
     "to_backend_service",
-)
-# nopycln: file
+)  # nopycln: file
