+++ services/web/server/src/simcore_service_webserver/catalog/models.py
@@ -1,0 +1,8 @@
+"""Catalog public facade for models per DESIGN.md §133-152."""
+
+from ._models import ServiceKeyVersionDict
+
+__all__: tuple[str, ...] = (
+    # models
+    "ServiceKeyVersionDict",
+)
