+++ services/web/server/src/simcore_service_webserver/chatbot/chatbot_service.py
@@ -1,7 +1,7 @@
 # mypy: disable-error-code=truthy-function
-from ._client import ChatbotRestClient, get_chatbot_rest_client
+from ._client import get_chatbot_rest_client
 
-__all__ = [
-    "ChatbotRestClient",
+__all__: tuple[str, ...] = (
+    # functions
     "get_chatbot_rest_client",
-]
+)
