+++ services/web/server/src/simcore_service_webserver/chatbot/models.py
@@ -1,0 +1,11 @@
+"""Chatbot public facade for models per DESIGN.md §133-152."""
+
+from ._client import ChatbotRestClient, ChatResponse, Message, ResponseItem, ResponseMessage
+
+__all__: tuple[str, ...] = (
+    "ChatResponse",
+    "ChatbotRestClient",
+    "Message",
+    "ResponseItem",
+    "ResponseMessage",
+)
